Refactor: main.js에서 api 요청부 분리
